Fix Bug

+++ app/Http/Controllers/EquiposController.php
@@ -3,17 +3,16 @@
 namespace App\Http\Controllers;
 
 @session_start();
-use Illuminate\Http\Request;
-use Aws\S3\S3Client;
 
 use App\Equipo;
+use Illuminate\Http\Request;
 
 class EquiposController extends Controller
 {
     public function index()
     {
-       $equipos=Equipo::paginate(25);
-        return view('equipos.index')->with('equipos',$equipos);
+        $equipos = Equipo::paginate(25);
+        return view('equipos.index')->with('equipos', $equipos);
     }
 
     public function create()
@@ -29,7 +28,7 @@ class EquiposController extends Controller
 
         try {
             $validator = \Validator::make($request->all(), $rules);
-            if ($validator->fails()){
+            if ($validator->fails()) {
                 return back()->withErrors($validator)->withInput();
             }
 
@@ -56,14 +55,14 @@ class EquiposController extends Controller
                 ));
             }*/
 
-            $equipo=Equipo::create([
+            $equipo = Equipo::create([
                 'nombre' => $request->nombre,
                 'bandera' => $fileName,
             ]);
-            return redirect()->route('equipos.edit', codifica($equipo->id))->with("notificacion","Se ha guardado correctamente su información");
+            return redirect()->route('equipos.edit', codifica($equipo->id))->with("notificacion", "Se ha guardado correctamente su información");
 
         } catch (Exception $e) {
-            \Log::info('Error creating item: '.$e);
+            \Log::info('Error creating item: ' . $e);
             return \Response::json(['created' => false], 500);
         }
     }
@@ -75,35 +74,37 @@ class EquiposController extends Controller
 
     public function edit($id)
     {
-        $id=decodifica($id);
-        $equipo=Equipo::find($id);
-        $_SESSION['equipo_id']=$id;
-        return view('equipos.edit')->with('equipo',$equipo);
+        $id = decodifica($id);
+        $equipo = Equipo::find($id);
+        $_SESSION['equipo_id'] = $id;
+        return view('equipos.edit')->with('equipo', $equipo);
     }
 
     public function update(Request $request, $id)
     {
         $rules = [
             'nombre' => 'required',
-            ];
+        ];
 
         try {
             $validator = \Validator::make($request->all(), $rules);
-            if ($validator->fails()){
+            if ($validator->fails()) {
                 return back()->withErrors($validator)->withInput();
             }
-            $id=decodifica($id);
+            $id = decodifica($id);
 
             $equipo = Equipo::where('id', $id)->first();
 
-            $data=[
+            $data = [
                 'nombre' => $request->nombre,
             ];
 
 
-            $this->deleteFile($equipo->bandera, 'equipos/');
-            $fileName = $this->saveFile($request->archivo, 'equipos/');
-            $data['bandera']=$fileName;
+            if ($request->archivo) {
+                $this->deleteFile($equipo->bandera, 'equipos/');
+                $fileName = $this->saveFile($request->archivo, 'equipos/');
+                $data['bandera'] = $fileName;
+            }
 
             $equipo->update($data);
 
@@ -130,24 +131,24 @@ class EquiposController extends Controller
                 $data['bandera']=$fileName;
             }
             Equipo::find($id)->update($data);*/
-            return redirect()->route('equipos.edit', codifica($id))->with("notificacion","Se ha guardado correctamente su información");
+            return redirect()->route('equipos.edit', codifica($id))->with("notificacion", "Se ha guardado correctamente su información");
 
         } catch (Exception $e) {
-            \Log::info('Error creating item: '.$e);
+            \Log::info('Error creating item: ' . $e);
             return \Response::json(['created' => false], 500);
         }
     }
 
     public function destroy($id)
     {
-        $id=decodifica($id);
-        try{
+        $id = decodifica($id);
+        try {
             $equipo = Equipo::where('id', $id)->first();
             $this->deleteFile($equipo->bandera, 'equipos/');
             $equipo->delete();
             return redirect()->route('equipos.index');
         } catch (\Illuminate\Database\QueryException $e) {
-            return back()->with("notificacion_error","Se ha producido un error, es probable que exista contenido relacionado a este registro que impide que se elimine");
+            return back()->with("notificacion_error", "Se ha producido un error, es probable que exista contenido relacionado a este registro que impide que se elimine");
         }
     }
 }
